libtracefs: Force off trace mmapping
The libtracefs API for trace buffer memory mapping should be stable. But the
Linux kernel has not yet accepted the changes. There's a chance that the kernel
API may change. In order to add the libtracefs API without depending on the API
being stable, add it but always have it fail unless the code is built with:

  make EXTRA_FLAGS=-DFORCE_MMAP_ENBALE  # For debugging purposes only!

Applications can start using the API and when it becomes available in the
kernel, force disabling will be removed in a "fix" update to libtracefs. This
way if the Linux kernel API changes before it gets in, an application using an
older libtracefs will not break, as the mapping code will still silently fail
like it always has. For an application to work with the buffer mmap after it is
in the Linux kernel with a stable API, all it will need to do is update the
libtracefs to a version that has it enabled.

That is, the libtracefs API of:

 tracefs_cpu_open_mapped()
 tracefs_cpu_is_mapped()
 tracefs_cpu_map()
 tracefs_cpu_unmap()

Will still be applicable even if the Linux kernel changes how it does the
memory mapping when it is finally upstream.

Link: https://lore.kernel.org/linux-trace-devel/20240108131903.52b02897@gandalf.local.home

Cc: Vincent Donnefort <vdonnefort@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>