libtracefs: Have tracefs_mmap_read() include subbuf meta data
The tracefs_cpu_read() returns tracefs_mmap_read() when the tcpu is mapped.
But tracefs_cpu_read() is supposed to return the amount that was read, which
includes the sub-buffer meta data. But kbuffer_read_buffer() only returns
the amount of data that was read and does not include the sub-buffer meta
data.

Link: https://lore.kernel.org/linux-trace-devel/20240110030116.81837-3-rostedt@goodmis.org

Cc: Vincent Donnefort <vdonnefort@google.com>
Fixes: 2ed14b59 ("libtracefs: Add ring buffer memory mapping APIs")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>