libtracefs: Call mmap ioctl if a refresh happens
If the reader sub-buffer gets more data on it due to a writer still on the
reader sub-buffer, just updating the kbuf is not enough. An ioctl() needs to
be called again to update the "read" pointer, otherwise after reading the
full reader sub-buffer and calling the next ioctl(), the kernel will not
swap out for a new reader sub-buffer as it still thinks there's more to be
read.

Link: https://lore.kernel.org/linux-trace-devel/20240110030116.81837-6-rostedt@goodmis.org

Cc: Vincent Donnefort <vdonnefort@google.com>
Fixes: 2ed14b594f669 ("libtracefs: Add ring buffer memory mapping APIs")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>