libtracefs: Fix tracefs_instance_reset() of triggers
The clearing of triggers may require a few loops as the order of the
clearing is important. Some triggers rely on other triggers to be cleared
before they can be cleared.

The tracefs_instance_reset() will loop over all the triggers, clearing them
one at a time, and restart from the top again if not all the triggers were
cleared. This guarantees that the triggers will have other triggers cleared
first if they depend on them to be cleared.

The "file" string used to clear the trigger wasn't updated to equal the next
file in the list, so it was stale.

Link: https://lore.kernel.org/linux-trace-devel/20240111173809.7d5462e8@gandalf.local.home

Fixes: 93e20af8 ("libtracefs: Fix tracefs_instance_reset to clear synthetic events")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>