libtracefs utest: Add better logic to cause missed events
When the system is running without much debug, it is possible to run the
function tracer without triggering missed events. As the tests are expecting
to have missed events to test the missed events handler, it gfails the test
because no missed events happen, and the missed events handler is correctly
not called. But the tests expect it to be called.

Add more logic to force missed events.

 * Shrink the buffer to just 4 pages.
 * Run "ls -l /usr/bin > /dev/null" instead of sleep().

Link: https://lore.kernel.org/linux-trace-devel/20240723220853.489058-4-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>