libtrace: Fix memory leak in tracefs_cpu
The introduction of the interface to allow the tracefs_cpu to handle buffer
allocation on their own caused a memory leak.  (Commit 1e1cc549d7
("libtracefs: Add API to read tracefs_cpu and return a kbuffer")).

The help of Address Sanitizer found the root cause. The "tcpu->buffer" was
being allocated but not freed. The fix will free memory during close and
cleanup functions.

Link: https://lore.kernel.org/AM9PR05MB77778BE58632548B4D3EEBECB5DF2@AM9PR05MB7777.eurprd05.prod.outlook.com
Signed-off-by: Philipp Kutzer <philipp.kutzer@siemens-healthineers.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>