libtracefs: Fix trace_sql() to handle buckets cast
The "CAST(start.delta as buckets=50)" doesn't work as the sqlhist.y doesn't
parse the =50 properly, and even if it did, the code in the histogram logic
fails to process the count.

Update the sqlhist.y (and run make sqlhist_remake) and fix the histogram
logic to process the buckets count.

Link: https://lore.kernel.org/20250410185124.7281b197@gandalf.local.home
Fixes: faa0ba7f ("libtracefs: Add buckets field to tracefs_sql() function")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>