libtracefs: Add trace_sql.bash for tracefs_sql() bash completions
trace-cmd and the internal sqlhist programs can take SQL input to be passed
into the tracefs_sql() function. This can be a bit complex, so create a
bash completion script that uses trace-cmd to allow for tab completions on a
bash command line to fill in the next commands.

This should simplify creating bash completions as well as make some
shortcuts known.

Other completions can include this file and use tracefs_sql_completion()
function.

Link: https://lore.kernel.org/20250410211516.1188030c@gandalf.local.home
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>