libtracefs: Fix tracefs_event_is_enabled() for all events
When tracefs_event_is_ensabled() is called with both system and event
as NULL in order to get the state of all events, it tries to disable all
events. This is becaues the call to enable_disable_all() doesn't take
the state as a parameter and just looks at the "enable" flag.

Link: https://lore.kernel.org/20250811175552.662cda17@batman.local.home
Fixes: fc94d1a8 ("libtracefs: Add tracefs_event_enable/disable() API")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>